% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts-add.R
\name{add_post}
\alias{add_post}
\title{Add a new bookmark}
\usage{
add_post(url, title, text = NULL, tags = NULL, timestamp = NULL,
  replace = NULL, shared = NULL, toread = NULL, auth_token = NULL)
}
\arguments{
\item{url}{URL for the bookmark (required)}

\item{title}{Title (required)}

\item{tags}{Up to 100 tags}

\item{timestamp}{If NULL, current time is used}

\item{replace}{Replace any existing bookmark with this URL. Default is TRUE. If set to FALSE, will throw an error if bookmark exists}

\item{shared}{Make bookmark public. Default is TRUE unless user has enabled the "save all bookmarks as private" user setting, in which case default is FALSE}

\item{toread}{Marks the bookmark as unread. Default is FALSE}

\item{auth_token}{Your auth_token}
}
\description{
Add a new bookmark
}
