% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts-all.R
\name{all_posts}
\alias{all_posts}
\title{Get all bookmarks}
\usage{
all_posts(tags = NULL, start = NULL, results = NULL, from = NULL,
  to = NULL, meta = TRUE)
}
\arguments{
\item{tags}{filter by up to three tags}

\item{start}{offset value (default is 0)}

\item{results}{number of results to return. Default is all}

\item{from}{return only bookmarks created after this time}

\item{to}{return only bookmarks created before this time}

\item{meta}{include a change detection signature for each bookmark}
}
\description{
Get all bookmarks
}
