% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zapi.R
\docType{data}
\name{api}
\alias{api}
\alias{posts}
\alias{tags}
\title{Use the Pinboard API}
\format{An object of class \code{list} of length 8.}
\usage{
posts

tags
}
\description{
\code{posts} and \code{tags} implement the associated endpoints in the
Pinboard API documentation (\url{https://pinboard.in/api/}). See the links
below for documentation of the individual functions.
}
\section{Posts}{

\itemize{
\item{\code{add}: \code{\link{add_post}}}
\item{\code{delete}: \code{\link{delete_post}}}
\item{\code{update}: \code{\link{last_updated}}}
\item{\code{all}: \code{\link{all_posts}}}
\item{\code{recent}: \code{\link{recent_posts}}}
\item{\code{get}}: \code{\link{get_posts}}
\item{\code{suggest}}: \code{\link{suggest}}
}
}

\section{Tags}{

\itemize{
\item{\code{get}: \code{\link{get_tags}}}
\item{\code{delete}: \code{\link{delete_tag}}}
\item{\code{rename}: \code{\link{rename_tag}}}
}
}

\examples{
\dontrun{
# to view all bookmarks with the "rstats" tag
posts$dates(tags = "rstats")
}

}
\keyword{datasets}
