% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{authentication}
\alias{authentication}
\title{How to setup your authentication}
\description{
All interactions with the 'pinboard.in' API (application programming interface)
require authentication. This doc explains where to find your secrets and how
to set up your computer so you can programmaticaly interact with pinboard.
}
\details{
You do need to have a pinboard account.
\subsection{Finding your secrets}{

Go to the \href{https://pinboard.in/settings/password}{pinboard password page}
and scroll down to API Token.
It says something like: "this is your API token:
\code{username:NUMBERSANDLETTERS}
}

\subsection{.Renvironment file}{

I would recommend you add the token to your .renviron file, either locally or
globally. A full introduction how these files work is beyond the scope
of this doc but if you create a 'hidden' file named .Renviron in the project folder
or in your home folder, R will read in the username and token and they will
be available to pinboardr.
Use \code{usethis::edit_r_environ()} or find the file manually.

Add to your .renviron file the following
\code{PB_USERNAME=username}
\code{PB_TOKEN="NUMBERSANDLETTERS"}

Restart the session to make the changes active.
}

\subsection{Setting variables locally}{

This is not the recommended approach, but will work nevertheless.
You can either:
\itemize{
\item set the variables at the start of your session using \code{Sys.setenv(PB_USERNAME = "username", PB_TOKEN="NUMBERSANDLETTERS")}
\item pass the username and token to every command: \code{pb_last_update(username=username, token=NUMBERSANDLETTERS)}
}
\subsection{Why is this not recommended?}{

This means your secrets, username and token are visible in your .rhistory, and if you save it in a
script, it will be visible to anyone who opens your script. I think you don't want that, but
my threatmodel is not your threatmodel.
}

}
}
