% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts-delete.R
\name{delete_post}
\alias{delete_post}
\title{Delete a bookmark}
\usage{
delete_post(url, auth_token = NULL)
}
\arguments{
\item{url}{The url of the bookmark to delete}

\item{auth_token}{Your auth_token}
}
\description{
Delete a bookmark
}
