% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-delete.R
\name{delete_tag}
\alias{delete_tag}
\title{Delete an existing tag}
\usage{
delete_tag(tag = NULL, auth_token = NULL)
}
\arguments{
\item{tag}{the tag to delete}

\item{auth_token}{Your auth_token}
}
\description{
Delete an existing tag
}
