% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts-get.R
\name{get_posts}
\alias{get_posts}
\title{Get posts from a given date}
\usage{
get_posts(tags = NULL, date = NULL, url = NULL, meta = TRUE,
  auth_token = NULL)
}
\arguments{
\item{tags}{character vector of up to 3 tags to filter by}

\item{date}{a \code{Date} to filter bookmarks}

\item{url}{Only return the bookmark for this URL}

\item{meta}{Include a change detection signature?}

\item{auth_token}{authentication token}
}
\description{
If no date or URL is given, the date of the most recent bookmark will be used.
}
