% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-get.R
\name{get_tags}
\alias{get_tags}
\title{Get a full list of tags along with the number of times they were used.}
\usage{
get_tags(auth_token = NULL)
}
\arguments{
\item{auth_token}{Your auth_token}
}
\description{
Get a full list of tags along with the number of times they were used.
}
