% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts-update.R
\name{last_updated}
\alias{last_updated}
\title{The most recent time a bookmark was added, updated or deleted}
\usage{
last_updated(auth_token = NULL)
}
\description{
Use this before calling posts/all to see if the data has changed since the last fetch.
}
\details{
@param auth_token Your auth_token

@export
}
