% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pb_add_tag_column}
\alias{pb_add_tag_column}
\title{Extract one tag from tag columns into separate column.}
\usage{
pb_add_tag_column(dataframe, tag)
}
\arguments{
\item{dataframe}{the dataframe of bookmarks}

\item{tag}{tag you would like to extract into a new column}
}
\value{
a dataframe with one extra column, containing TRUE or FALSE for presence of the tag
}
\description{
Convenience function to make working with the dataframe of bookmarks
a bit easier.
}
\examples{
\dontrun{
pb_add_tag_column(all_bookmars, "europe")
}
}
