% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{pb_get_note}
\alias{pb_get_note}
\title{Get one specific note}
\usage{
pb_get_note(id = NULL, username = NULL, token = NULL)
}
\arguments{
\item{id}{the id of a note}

\item{username}{your username You probably don't need to supply this every time. See \code{?authentication}}

\item{token}{your token}
}
\value{
a list containing id, title, created_at, updated_at, length, text and hash
}
\description{
Get one specific note
}
\examples{
\dontrun{
pb_get_note(id=12345)
}
}
