% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{pb_get_notes_overview}
\alias{pb_get_notes_overview}
\title{Returns a data.frame of the user's notes}
\usage{
pb_get_notes_overview(username = NULL, token = NULL)
}
\arguments{
\item{username}{your username You probably don't need to supply this every time. See \code{?authentication}}

\item{token}{your token}
}
\value{
a data.frame with notes meta information
}
\description{
Return a data.frame of notes meta information,
id, hash, title, length, created_at and updated_at.
When you know the id, you can use the other endpoint
pb_get_note
}
