% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{pb_posts_add}
\alias{pb_posts_add}
\title{Add a bookmark}
\usage{
pb_posts_add(
  url,
  title,
  description = NULL,
  tags = NULL,
  dt = NULL,
  replace = NULL,
  public = NULL,
  toread = NULL,
  username = NULL,
  token = NULL
)
}
\arguments{
\item{url}{\strong{REQUIRED} the URL of the item}

\item{title}{\strong{REQUIRED} Title of the item.}

\item{description}{Description of the item}

\item{tags}{vector of up to 100 tags}

\item{dt}{creation time for this bookmark. Defaults to current time. Datestamps more than 10 minutes ahead of server time will be reset to current server time (UTC timestamp in this format: 2010-12-11T19:48:02Z)}

\item{replace}{Replace any existing bookmark with this URL. Default is TRUE If set to FALSE, will throw an error if bookmark exists}

\item{public}{TRUE/FALSE    Make bookmark public. Default is "TRUE" unless user has enabled the "save all bookmarks as private" user setting, in which case default is "no"}

\item{toread}{TRUE/FALSE    Marks the bookmark as unread. Default is "FALSE"}

\item{username}{your username You probably don't need to supply this every time. See \code{?authentication}}

\item{token}{your token}
}
\value{
status in text 'done'
}
\description{
You must give a url and description, all other arguments are optional.
}
\examples{
\dontrun{
pb_posts_add(url="https://example.com", title="an example website")
}
}
