% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{pb_posts_all}
\alias{pb_posts_all}
\title{Return all bookmarks in the user's account}
\usage{
pb_posts_all(
  tags = NULL,
  start = 0L,
  results = NULL,
  fromdt = NULL,
  todt = NULL,
  meta = NULL,
  username = NULL,
  token = NULL
)
}
\arguments{
\item{tags}{filter by up to three tags}

\item{start}{int    offset value (default is 0)}

\item{results}{int    number of results to return. Default is all}

\item{fromdt}{datetime    return only bookmarks created after this time}

\item{todt}{datetime    return only bookmarks created before this time}

\item{meta}{int    include a change detection signature for each bookmark}

\item{username}{your username You probably don't need to supply this every time. See \code{?authentication}}

\item{token}{your token}
}
\value{
a dataframe of all posts of pinboard format see \code{?pinboard_formatting}
}
\description{
This API endpoint is rate limited to once every 5 minutes.
}
\examples{
\dontrun{
pb_posts_all(tags = "inspiration", results = 15)
}
}
