% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{pb_posts_dates}
\alias{pb_posts_dates}
\title{Posts per date}
\usage{
pb_posts_dates(tags = NULL, username = NULL, token = NULL)
}
\arguments{
\item{tags}{filter by up to three tags}

\item{username}{your username You probably don't need to supply this every time. See \code{?authentication}}

\item{token}{your token}
}
\value{
a data.frame with dates and counts
}
\description{
Returns a dataframe of dates with the number of posts at each date.
Use this to see all posts per date or posts per tag per date.
}
\examples{
\dontrun{
pb_posts_dates(tags = "europe")
}
}
