% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{pb_posts_delete}
\alias{pb_posts_delete}
\title{Delete a bookmark}
\usage{
pb_posts_delete(url = NULL, username = NULL, token = NULL)
}
\arguments{
\item{url}{\strong{REQUIRED} url of bookmark to delete}

\item{username}{your username You probably don't need to supply this every time. See \code{?authentication}}

\item{token}{your token}
}
\value{
status in text: "done", "item not found"
}
\description{
Delete a bookmark from pinboard by url.
}
\examples{
\dontrun{
pb_posts_delete("https://example.com")
}
}
