% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{pb_posts_get}
\alias{pb_posts_get}
\title{Get all posts on a day}
\usage{
pb_posts_get(
  tags = NULL,
  dt = NULL,
  url = NULL,
  meta = NULL,
  username = NULL,
  token = NULL
)
}
\arguments{
\item{tags}{filter by up to three tags}

\item{dt}{return results bookmarked on this day (UTC date in this format: 2010-12-31).}

\item{url}{return bookmark for this URL}

\item{meta}{TRUE/FALSE    include a change detection signature in a meta attribute}

\item{username}{your username You probably don't need to supply this every time. See \code{?authentication}}

\item{token}{your token}
}
\value{
a dataframe of all posts of pinboard format see \code{?pinboard_formatting}
}
\description{
Returns one or more posts on a single day matching the arguments.
If no date or url is given, date of most recent bookmark will be used.
}
\examples{
\dontrun{
pb_posts_get()
}
}
