% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{pb_posts_recent}
\alias{pb_posts_recent}
\title{Get most recent posts}
\usage{
pb_posts_recent(tags = NULL, count = 15, username = NULL, token = NULL)
}
\arguments{
\item{tags}{tags filter by up to three tags}

\item{count}{number of results to return default is 15 max is 100}

\item{username}{your username You probably don't need to supply this every time. See \code{?authentication}}

\item{token}{your token}
}
\value{
a dataframe of all recent of pinboard format see \code{?pinboard_formatting}
}
\description{
Returns a list of the user's most recent posts, filtered by tag.
}
\examples{
\dontrun{
pb_posts_recent(tags = "inspiration")
}
}
