% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{pb_posts_suggest}
\alias{pb_posts_suggest}
\title{Returns a list of popular tags and recommended tags for a given URL.}
\usage{
pb_posts_suggest(url = NULL, username = NULL, token = NULL)
}
\arguments{
\item{url}{\strong{REQUIRED} url to get suggestions for}

\item{username}{your username You probably don't need to supply this every time. See \code{?authentication}}

\item{token}{your token}
}
\value{
a data.frame with popular and recommended tags
}
\description{
Popular tags are tags used site-wide for the url; recommended tags are drawn from the user's own tags.
}
\examples{
\dontrun{
pb_posts_suggest(url = "https://example.com")
}
}
