% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{pb_tags_delete}
\alias{pb_tags_delete}
\title{Delete an existing tag}
\usage{
pb_tags_delete(tag, username = NULL, token = NULL)
}
\arguments{
\item{tag}{\strong{REQUIRED} tag the delete}

\item{username}{your username You probably don't need to supply this every time. See \code{?authentication}}

\item{token}{your token}
}
\value{
text confirmation
}
\description{
Notice that 'pinboard.in' does not tell you if a tag
exists or not. It will return a 'done' unless something else goes wrong.
}
\examples{
\dontrun{
pb_tags_delete(tag = "svg")
}
}
