% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{pb_tags_get}
\alias{pb_tags_get}
\title{Find all tags in use.}
\usage{
pb_tags_get(username = NULL, token = NULL)
}
\arguments{
\item{username}{your username You probably don't need to supply this every time. See \code{?authentication}}

\item{token}{your token}
}
\value{
a dataframe with all tags and their count
}
\description{
Find all tags in use.
}
\examples{
\dontrun{
pb_get_all_tags()
}
}
