% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{pb_tags_rename}
\alias{pb_tags_rename}
\title{Rename a tag}
\usage{
pb_tags_rename(old, new = NULL, username = NULL, token = NULL)
}
\arguments{
\item{old}{\strong{REQUIRED} old tag name}

\item{new}{\strong{REQUIRED} new tag name, if empty nothing will happen}

\item{username}{your username You probably don't need to supply this every time. See \code{?authentication}}

\item{token}{your token}
}
\value{
text confirmation
}
\description{
Change the name of a tag. Notice that 'pinboard.in' does not tell you if a tag
exists or not. It will return a 'done' unless something else goes wrong.
}
\examples{
\dontrun{
pb_tags_rename(old = "svg", new = "illustrations")
}
}
