% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{pb_user_api_token}
\alias{pb_user_api_token}
\title{Get your api token}
\usage{
pb_user_api_token(username = NULL, token = NULL)
}
\arguments{
\item{username}{your username You probably don't need to supply this every time. See \code{?authentication}}

\item{token}{your token}
}
\value{
token in string format
}
\description{
This is a silly endpoint and function, but I implemented it
anyways. To reach this endpoint, you already are in possession of your
token. And I will not implement a username and password authentication.
}
