% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{pinboard_formatting}
\alias{pinboard_formatting}
\title{Pinboard format data.frame}
\description{
The pinboardr package returns the data in a format as close as possible
to the data supplied by the API. Because 'pinboard.in' modeled their
API on the original 'Delicious API' with the same unfortunate
choices. See also the description on pinboard
\href{https://pinboard.in/api/#posts_add}{posts_add docs}.
}
\details{
I try to hide the complexity from you by translating but sometimes that
gives us different names. I first give you the pinboardr name
and than the name in the delicious and pinboard API:
\itemize{
\item The 'title' of a bookmark is actually called 'description'.
\item The 'description' of a bookmark is actually called 'extended'.
\item The 'public' setting of bookmarks is called 'shared'.
}
}
