% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinboardr-package.R
\docType{package}
\name{pinboardr-package}
\alias{pinboardr}
\alias{pinboardr-package}
\title{pinboardr: Connect to 'pinboard.in' service}
\description{
Full connectivity to the 'pinboard.in' service. Retrieve bookmarks, add new ones. 
    Add, delete and rename tags, all from the convenience of your R session.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RMHogervorst/pinboardr}
  \item Report bugs at \url{https://github.com/RMHogervorst/pinboardr/issues}
}

}
\author{
\strong{Maintainer}: Roel M. Hogervorst \email{hogervorst.rm@gmail.com} (\href{https://orcid.org/0000-0001-7509-0328}{ORCID})

}
\keyword{internal}
