% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts-dates.R
\name{post_dates}
\alias{post_dates}
\title{Get a list of dates with the number of posts at each date.}
\usage{
post_dates(tags, auth_token = NULL)
}
\arguments{
\item{tags}{filter by up to three tags}

\item{auth_token}{Your auth_token}
}
\description{
Get a list of dates with the number of posts at each date.
}
