% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts-recent.R
\name{recent_posts}
\alias{recent_posts}
\title{Get recent posts}
\usage{
recent_posts(tags = NULL, count = NULL, auth_token = NULL)
}
\arguments{
\item{tags}{Up to 3 tags to filter by}

\item{count}{How many posts to retrieve. Default is 15, maximum is 100}

\item{auth_token}{Your auth_token}
}
\description{
Get recent posts
}
