% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-rename.R
\name{rename_tag}
\alias{rename_tag}
\title{Rename a tag}
\usage{
rename_tag(old = NULL, new = NULL, auth_token = NULL)
}
\arguments{
\item{old}{Tag to rename}

\item{new}{Replacement tag}

\item{auth_token}{Your auth_token}
}
\description{
Rename a tag
}
