% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts-suggest.R
\name{suggest}
\alias{suggest}
\title{List of popular tags and recommended tags for a given URL.}
\usage{
suggest(url, auth_token = NULL)
}
\arguments{
\item{url}{A URL}

\item{auth_token}{Your auth_token

 @export}
}
\description{
Popular tags are tags used site-wide for the url; recommended tags are drawn from
the user's own tags.
}
